bpf: Fixes to support mattr=+alu32
[ upstream commit ad0d3cf341402feaf34cb612c0f57732b0f5d5b4 ]

mattr=+alu32, supported since LLVM 7.0 and implied by mcpu=v3, enables
the use of 32-bit registers in BPF bytecode. Enabling this compiler
option can however result in loading issues as illustrated below.

    12: (61) r1 = *(u32 *)(r0 +80) // ctx->data_end
    13: (61) r6 = *(u32 *)(r0 +76) // ctx->data
    14: (bc) w7 = w6 // <- verifier looses track of inferred pkt type here.
    [...]
    38: (71) r1 = *(u8 *)(r7 +20)
    R7 invalid mem access 'inv'

These errors typically happen because the data and data_end pointers are
actually 32-bit registers. Depending on how these pointers are used,
LLVM sometimes makes use of that assumption (e.g., 32-bit assignment on
instruction 14 above). The verifier is however not able to follow and
reject such programs.

We can usually work around those by ensuring these pointers are only
used via 64-bit types. This commit implements this wherever needed to
pass the verifier.

+++ bpf/lib/metrics.h
@@ -21,7 +21,7 @@
  *		is the drop error code.
  * Update the metrics map.
  */
-static __always_inline void update_metrics(__u32 bytes, __u8 direction,
+static __always_inline void update_metrics(__u64 bytes, __u8 direction,
 					   __u8 reason)
 {
 	struct metrics_value *entry, newEntry = {};
@@ -34,10 +34,10 @@ static __always_inline void update_metrics(__u32 bytes, __u8 direction,
 	entry = map_lookup_elem(&METRICS_MAP, &key);
 	if (entry) {
 		entry->count += 1;
-		entry->bytes += (__u64)bytes;
+		entry->bytes += bytes;
 	} else {
 		newEntry.count = 1;
-		newEntry.bytes = (__u64)bytes;
+		newEntry.bytes = bytes;
 		map_update_elem(&METRICS_MAP, &key, &newEntry, 0);
 	}
 }
